+++ src/main/java/de/maxhenkel/plane/ServerConfig.java
@@ -15,6 +15,7 @@ import java.util.stream.Collectors;
 public class ServerConfig extends ConfigBase {
 
     public final ModConfigSpec.DoubleValue maxPlaneDamage;
+    public final ModConfigSpec.BooleanValue planeToPlaneCollision;
 
     private final ModConfigSpec.ConfigValue<List<? extends String>> validFuelsSpec;
     public List<Tag<Fluid>> validFuels = new ArrayList<>();
@@ -22,6 +23,7 @@ public class ServerConfig extends ConfigBase {
     public ServerConfig(ModConfigSpec.Builder builder) {
         super(builder);
         maxPlaneDamage = builder.defineInRange("plane.max_damage", 200D, 100D, Double.MAX_VALUE);
+        planeToPlaneCollision = builder.define("plane_to_plane_collision", true);
 
         validFuelsSpec = builder.defineList("plane.valid_fuels", () -> List.of("#car:gas_station"), () -> "", Objects::nonNull);
     }